+++ src/SinkaPlugin.cs
@@ -28,6 +28,7 @@ namespace Sinka
 
             _harmony = new Harmony(PluginGuid);
             _harmony.PatchAll(typeof(ScenePatches));
+            _harmony.PatchAll(typeof(TorchPoles));
 
             Log.LogInfo(PluginName + " " + PluginVersion + " by " + PluginAuthor + " - ready.");
         }